+++ Navigation/Post.swift
@@ -1,0 +1,12 @@
+//
+//  Post.swift
+//  Navigation
+//
+//  Created by Is Re on 24/06/23.
+//
+
+import Foundation
+
+struct Post {
+    let title: String
+}
