+++ Navigation/PostViewController.swift
@@ -1,0 +1,33 @@
+//
+//  PostViewController.swift
+//  Navigation
+//
+//  Created by Is Re on 24/06/23.
+//
+
+import UIKit
+
+class PostViewController: UIViewController {
+
+    var post: Post?
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        view.backgroundColor = .systemBackground
+
+        if let post = post {
+            title = post.title
+        }
+        
+        let infoButton = UIBarButtonItem(image: UIImage(systemName: "info.circle"), style: .plain, target: self, action: #selector(showInfo))
+        navigationItem.rightBarButtonItem = infoButton
+    }
+
+    @objc private func showInfo() {
+        let infoViewController = InfoViewController()
+        let navigationController = UINavigationController(rootViewController: infoViewController)
+        navigationController.modalPresentationStyle = .fullScreen
+        present(navigationController, animated: true, completion: nil)
+    }
+}
