+++ Navigation/SwiftUIView.swift
@@ -1,0 +1,20 @@
+//
+//  SwiftUIView.swift
+//  Navigation
+//
+//  Created by Is Re on 24/06/23.
+//
+
+import SwiftUI
+
+struct SwiftUIView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct SwiftUIView_Previews: PreviewProvider {
+    static var previews: some View {
+        SwiftUIView()
+    }
+}
